obs-vst: Use v2 of obs_properties_add_button
v1 of obs_properties_add_button will be deprecated soon.
Also fixes UNUSED_PARAMETER calls in the callbacks for parameters that
are not actually unused, and moves the other ones to the top (as in most
of the rest of the project).

(cherry picked from commit cb75098a93aa04773ef0726037c07cf5d515d8ce)